requirements.txt
# Core Dependencies
aiohttp>=3.9.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Blockchain & Web3
web3>=6.11.0

# Data Analysis
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Configuration
PyYAML>=6.0

# Notifications
discord-webhook>=1.3.0
python-telegram-bot>=20.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.7.0
types-PyYAML>=6.0.0
